DatabaseBridge/requirements.txt
paho-mqtt == 2.1.0
psycopg2 == 2.9.10
influxdb-client == 1.47.0